Add support for "encrpyted" fortune files
Classic fortune uses the same file format for all fortune files.
However there are fortune files which might be found to be offensive.
To avoid having offensive words in plain sight the fortunes in these files
were "encrypted" using rot13.
If a file is to be considered offensive is decided by its creator and expressed
by adding "-o" as suffix to the file name. The suffix is also used as a trigger
whether the fortune content should be passed through rot13.